+++ src/game/verbs/verb.test.ts
@@ -13,7 +13,7 @@ import { Room } from "../items/room";
 import { checkAutoActions } from "../input/autoActionExecutor";
 import { ActionChain } from "../../utils/actionChain";
 import { AnyAction } from "redux";
-import { Effect, VerbRelation } from "../../utils/effects";
+import { Effect, EffectBuilder, VerbRelation } from "../../utils/effects";
 
 jest.mock("../input/autoActionExecutor", () => ({
   checkAutoActions: jest.fn(async () => true)
@@ -353,10 +353,7 @@ describe("chainable actions", () => {
 
   it("gives a default message if a prepositional verb fails and the player doesn't have a holdable indirect item", async () => {
     // The auto actions are mocked to return true here
-    const verb = new Verb.Builder("squeeze")
-      .withTest(false, "won't happen")
-      .makePrepositional("with what")
-      .build();
+    const verb = new Verb.Builder("squeeze").withTest(false, "won't happen").makePrepositional("with what").build();
     const plushie = new Item.Builder("plushie").isHoldable().build();
     selectInventory().addItem(plushie);
     verb.attempt(plushie, new Item.Builder("tongs").isHoldable().build());
@@ -499,7 +496,8 @@ describe("smart tests", () => {
 describe("effects", () => {
   let egg: Item,
     bat: Item,
-    verbSuccess = true;
+    verbSuccess = true,
+    builder: EffectBuilder;
 
   beforeEach(() => {
     unregisterStore();
@@ -520,18 +518,17 @@ describe("effects", () => {
     bat = new Item.Builder("bat").isHoldable().build();
     moveItem(egg, selectInventory());
     moveItem(bat, selectInventory());
+
+    builder = new Effect.Builder()
+      .withPrimaryItem("egg")
+      .withSecondaryItem("bat")
+      .withVerbName("hit")
+      .isSuccessful(true)
+      .withVerbRelation(VerbRelation.Instead);
   });
 
   it("triggers effects, after actions, objects available in actions", async () => {
-    addEffect(
-      new Effect.Builder()
-        .withPrimaryItem("egg")
-        .withSecondaryItem("bat")
-        .withVerbName("hit")
-        .isSuccessful(true)
-        .withVerbRelation(VerbRelation.Instead)
-        .withActions(({ item, other }) => `${other!.name} smash ${item.name}`)
-    );
+    addEffect(builder.withActions(({ item, other }) => `${other!.name} smash ${item.name}`));
     await egg.try("hit", bat);
     expect(selectCurrentPage()).toInclude("bat smash egg");
     expect(selectCurrentPage()).not.toInclude("hit it good");
@@ -540,15 +537,7 @@ describe("effects", () => {
   });
 
   it("optionally continues executing the verb after a successful effect", async () => {
-    addEffect(
-      new Effect.Builder()
-        .withPrimaryItem("egg")
-        .withSecondaryItem("bat")
-        .withVerbName("hit")
-        .isSuccessful(true)
-        .withVerbRelation(VerbRelation.Before)
-        .withActions(({ item, other }) => `${other!.name} smash ${item.name}`)
-    );
+    addEffect(builder.withVerbRelation(VerbRelation.Before).withActions(({ item, other }) => `${other!.name} smash ${item.name}`));
     await egg.try("hit", bat);
     expect(selectCurrentPage()).toInclude("bat smash egg");
     expect(selectCurrentPage()).toInclude("hit it good");
@@ -556,10 +545,7 @@ describe("effects", () => {
 
   it("optionally continues executing the verb after a failed effect", async () => {
     addEffect(
-      new Effect.Builder()
-        .withPrimaryItem("egg")
-        .withSecondaryItem("bat")
-        .withVerbName("hit")
+      builder
         .isSuccessful(false)
         .withVerbRelation(VerbRelation.Before)
         .withActions(({ item, other }) => `${other!.name} brush ${item.name}`)
@@ -570,15 +556,7 @@ describe("effects", () => {
   });
 
   it("triggers wildcard effects, objects available in actions", async () => {
-    addEffect(
-      new Effect.Builder()
-        .withAnyPrimaryItem()
-        .withSecondaryItem("bat")
-        .withVerbName("hit")
-        .isSuccessful(true)
-        .withVerbRelation(VerbRelation.Instead)
-        .withActions(({ item, other }) => `${other!.name} smash ${item.name}`)
-    );
+    addEffect(builder.withAnyPrimaryItem().withActions(({ item, other }) => `${other!.name} smash ${item.name}`));
     await egg.try("hit", bat);
     expect(selectCurrentPage()).toInclude("bat smash egg");
     expect(selectCurrentPage()).not.toInclude("hit it good");
@@ -587,14 +565,7 @@ describe("effects", () => {
   });
 
   it("optionally continues executing the verb after a successful wildcard effect", async () => {
-    addEffect(
-      new Effect.Builder()
-        .withAnyPrimaryItem()
-        .withSecondaryItem("bat")
-        .withVerbName("hit")
-        .withVerbRelation(VerbRelation.Before)
-        .withActions(({ item, other }) => `${other!.name} smash ${item.name}`)
-    );
+    addEffect(builder.withAnyPrimaryItem().withVerbRelation(VerbRelation.Before).withActions(({ item, other }) => `${other!.name} smash ${item.name}`));
     await egg.try("hit", bat);
     expect(selectCurrentPage()).toInclude("bat smash egg");
     expect(selectCurrentPage()).toInclude("hit it good");
@@ -602,10 +573,8 @@ describe("effects", () => {
 
   it("optionally continues executing the verb after a failed wildcard effect", async () => {
     addEffect(
-      new Effect.Builder()
+      builder
         .withAnyPrimaryItem()
-        .withSecondaryItem("bat")
-        .withVerbName("hit")
         .isSuccessful(false)
         .withVerbRelation(VerbRelation.Before)
         .withActions(({ item, other }) => `${other!.name} brush ${item.name}`)
@@ -617,14 +586,7 @@ describe("effects", () => {
 
   it("executes effects for normally non-prepositional verbs", async () => {
     const microscope = new Item.Builder("microscope").build();
-    addEffect(
-      new Effect.Builder()
-        .withPrimaryItem("egg")
-        .withSecondaryItem("microscope")
-        .withVerbName("examine")
-        .withVerbRelation(VerbRelation.Instead)
-        .withActions("It looks interesting")
-    );
+    addEffect(builder.withSecondaryItem("microscope").withVerbName("examine").withActions("It looks interesting"));
     await egg.try("examine", microscope);
     expect(selectCurrentPage()).toInclude("It looks interesting");
   });
@@ -632,14 +594,7 @@ describe("effects", () => {
   it("executes effects before the verb", async () => {
     // Check it's effect then verb, not verb then effect.
     let x = 60;
-    addEffect(
-      new Effect.Builder()
-        .withPrimaryItem("egg")
-        .withSecondaryItem("bat")
-        .withVerbName("hit")
-        .withVerbRelation(VerbRelation.Before)
-        .withActions(() => (x /= 3))
-    );
+    addEffect(builder.withVerbRelation(VerbRelation.Before).withActions(() => (x /= 3)));
     egg.verbs.hit.onSuccess.addAction(() => (x += 10));
     await egg.try("hit", bat);
     expect(x).toBe(30);
@@ -648,32 +603,51 @@ describe("effects", () => {
   it("executes effects after the verb", async () => {
     // Check it's verb then effect, not effect then verb.
     let x = 60;
-    addEffect(
-      new Effect.Builder()
-        .withPrimaryItem("egg")
-        .withSecondaryItem("bat")
-        .withVerbName("hit")
-        .withVerbRelation(VerbRelation.After)
-        .withActions(() => (x /= 2))
-    );
+    addEffect(builder.withVerbRelation(VerbRelation.After).withActions(() => (x /= 2)));
     egg.verbs.hit.onSuccess.addAction(() => (x += 10));
     await egg.try("hit", bat);
     expect(x).toBe(35);
   });
 
-  it("doesn't execute after the verb if the verb is unsuccessful", async () => {
+  it("doesn't execute an effect after the verb if the verb is unsuccessful", async () => {
     verbSuccess = false;
     let x = 60;
-    addEffect(
-      new Effect.Builder()
-        .withPrimaryItem("egg")
-        .withSecondaryItem("bat")
-        .withVerbName("hit")
-        .withVerbRelation(VerbRelation.After)
-        .withActions(() => (x /= 2))
-    );
+    addEffect(builder.withVerbRelation(VerbRelation.After).withActions(() => (x /= 2)));
     egg.verbs.hit.onSuccess.addAction(() => (x += 10));
     await egg.try("hit", bat);
     expect(x).toBe(60);
+  });
+
+  it("doesn't execute an effect if its tests fail", async () => {
+    let x = 0;
+    addEffect(builder.withTest(() => 2 > 3, "fail").withActions(() => x++));
+    await egg.try("hit", bat);
+    expect(x).toBe(0);
+    expect(selectCurrentPage()).toInclude("fail");
+  });
+
+  it("executes an effect if its tests pass", async () => {
+    let x = 0;
+    addEffect(builder.withTest(() => 2 > 1).withActions(() => x++));
+    await egg.try("hit", bat);
+    expect(x).toBe(1);
+  });
+
+  it("isn't considered successful if the effect's tests fail", async () => {
+    addEffect(builder.withVerbRelation(VerbRelation.Before).withTest(() => 2 < 1).withActions("nay"));
+    verbSuccess = false; // Neither effect nor verb tests pass, so verb fails.
+    await egg.try("hit", bat);
+    expect(selectCurrentPage()).toInclude("missed it");
+  });
+
+  it("may have multiple tests", async () => {
+    let x = 10;
+    addEffect(builder.withTest(() => 2 > 1).withTest(() => x === 0, "not zero").withTest(true).withActions(() => x++));
+    await egg.try("hit", bat);
+    expect(x).toBe(10);
+    expect(selectCurrentPage()).toInclude("not zero");
+    x = 0;
+    await egg.try("hit", bat);
+    expect(x).toBe(1);
   });
 });